rag-assistant/requirements.txt
# Core RAG Dependencies
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2

# Vector Store
faiss-cpu==1.12.0
chromadb==0.6.3

# Embeddings
sentence-transformers==5.2.2
tokenizers==0.22.2

# Document Loaders
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.25

# Text Processing
tiktoken==0.12.0
nltk==3.9.1

# LLM Integration
openai==1.58.1
anthropic==0.42.0

# UI
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
rich==13.9.4
tqdm==4.67.1
numpy==1.26.4
